Refactor map structure
