+++ src/components/cardFilters/EffectFilter.js
