+++ src/components/cardFilters/StrengthFilter.js
