+++ src/pages/auth/SignIn.js
@@ -2,19 +2,14 @@ import { useState } from 'react'
 import { connect } from 'react-redux'
 import { Link } from 'react-router-dom'
 
-import { signUp } from '../redux/auth/actions'
+import { signIn } from '../../redux/auth/actions'
 
-function SignUpUI(props) {
+function SignInUI(props) {
   const [username, setUsername] = useState('')
-  const [email, setEmail] = useState('')
   const [password, setPassword] = useState('')
 
   const handleUsernameChange = (e) => {
     setUsername(e.target.value)
-  }
-
-  const handleEmailChange = (e) => {
-    setEmail(e.target.value)
   }
 
   const handlePasswordChange = (e) => {
@@ -22,7 +17,7 @@ function SignUpUI(props) {
   }
 
   const handleLogin = () => {
-    props.signUp(username, email, password)
+    props.signIn(username, password)
   }
 
   return (
@@ -42,23 +37,8 @@ function SignUpUI(props) {
             className="shadow appearance-none border rounded w-full py-2 px-3"
             id="username"
             type="text"
+            placeholder="Fill in username or email..."
             onChange={handleUsernameChange}
-          />
-        </div>
-
-        <div className="mb-4">
-          <label
-            className="text-left ml-5 block text-gray-500 text-sm font-bold mb-2"
-            htmlFor="email"
-          >
-            Email
-          </label>
-
-          <input
-            className="shadow appearance-none border rounded w-full py-2 px-3"
-            id="email"
-            type="email"
-            onChange={handleEmailChange}
           />
         </div>
 
@@ -88,10 +68,10 @@ function SignUpUI(props) {
           </button>
 
           <Link
-            to="/signin"
+            to="/signup"
             className="inline-block align-baseline font-bold text-sm text-blue-500 hover:text-blue-800"
           >
-            Have an account??
+            No account?
           </Link>
         </div>
       </div>
@@ -100,8 +80,7 @@ function SignUpUI(props) {
 }
 
 const mapDispatchToProps = (dispatch) => ({
-  signUp: (username, email, password) =>
-    dispatch(signUp(username, email, password)),
+  signIn: (username, password) => dispatch(signIn(username, password)),
 })
 
-export const SignUp = connect(null, mapDispatchToProps)(SignUpUI)
+export const SignIn = connect(null, mapDispatchToProps)(SignInUI)
